master: apply portable game fixes

+++ src/BASE/RESMGR.cpp
@@ -345,8 +345,8 @@ void resourceManager::PointToFile(u32l fileId) {
         sprintf(
             gText,
             "ResMgr::PointToFile failure!  ThisFileId:%d  LastFileId:%d  LastFileName:%s",
-            fileId,
-            m_lastFileId,
+            static_cast<i32>(fileId),
+            static_cast<i32>(m_lastFileId),
             m_lastFileName
         );
         ShutDown(gText);
@@ -379,8 +379,8 @@ u32l resourceManager::GetFileSize(u32l fileId) {
         sprintf(
             gText,
             "ResMgr::PointToFile failure!  ThisFileId:%d  LastFileId:%d  LastFileName:%s",
-            fileId,
-            m_lastFileId,
+            static_cast<i32>(fileId),
+            static_cast<i32>(m_lastFileId),
             m_lastFileName
         );
         ShutDown(gText);
@@ -448,7 +448,7 @@ void resourceManager::ReadBlock(i8* destination, u32l size) {
             gText,
             "File error - bytes read %d, bytes requested %d, errno %d, last file '%s'",
             bytesRead,
-            size,
+            static_cast<i32>(size),
             errno,
             m_lastFileName
         );
